Struct surf::RequestBuilder

pub struct RequestBuilder { /* private fields */ }

Request Builder

Provides an ergonomic way to chain the creation of a request. This is generally accessed as the return value from surf::{method}(), however Request::builder is also provided.

Examples

use surf::http::{Method, mime::HTML, Url};
let mut request = surf::post("https://httpbin.org/post")
    .body("<html>hi</html>")
    .header("custom-header", "value")
    .content_type(HTML)
    .build();

assert_eq!(request.take_body().into_string().await.unwrap(), "<html>hi</html>");
assert_eq!(request.method(), Method::Post);
assert_eq!(request.url(), &Url::parse("https://httpbin.org/post")?);
assert_eq!(request["custom-header"], "value");
assert_eq!(request["content-type"], "text/html;charset=utf-8");

use surf::http::{Method, Url};
let url = Url::parse("https://httpbin.org/post")?;
let request = surf::Request::builder(Method::Post, url).build();

Implementations

impl RequestBuilder

pub fn new(method: Method, url: Url) -> Self

Create a new instance.

This method is particularly useful when input URLs might be passed by third parties, and you don’t want to panic if they’re malformed. If URLs are statically encoded, it might be easier to use one of the shorthand methods instead.

Examples

use surf::http::{Method, Url};

let url = Url::parse("https://httpbin.org/get")?;
let req = surf::RequestBuilder::new(Method::Get, url).build();

pub fn header(
self,
key: impl Into<HeaderName>,
value: impl ToHeaderValues
) -> Self

Sets a header on the request.

Examples

let req = surf::get("https://httpbin.org/get").header("header-name", "header-value").build();
assert_eq!(req["header-name"], "header-value");

pub fn content_type(self, content_type: impl Into<Mime>) -> Self

Sets the Content-Type header on the request.

Examples

let req = surf::post("https://httpbin.org/post").content_type(mime::HTML).build();
assert_eq!(req["content-type"], "text/html;charset=utf-8");

pub fn body(self, body: impl Into<Body>) -> Self

Sets the body of the request from any type with implements Into<Body>, for example, any type with is AsyncRead.

Mime

The encoding is set to application/octet-stream.

Examples

use serde_json::json;
let mut req = surf::post("https://httpbin.org/post").body(json!({ "any": "Into<Body>"})).build();
assert_eq!(req.take_body().into_string().await.unwrap(), "{\"any\":\"Into<Body>\"}");

pub fn body_json(self, json: &impl Serialize) -> Result<Self>

Pass JSON as the request body.

Mime

The encoding is set to application/json.

Errors

This method will return an error if the provided data could not be serialized to JSON.

Examples

#[derive(Deserialize, Serialize)]
struct Ip {
    ip: String
}

let uri = "https://httpbin.org/post";
let data = &Ip { ip: "129.0.0.1".into() };
let res = surf::post(uri).body_json(data)?.await?;
assert_eq!(res.status(), 200);

pub fn body_string(self, string: String) -> Self

Pass a string as the request body.

Mime

The encoding is set to text/plain; charset=utf-8.

Examples

let uri = "https://httpbin.org/post";
let data = "hello world".to_string();
let res = surf::post(uri).body_string(data).await?;
assert_eq!(res.status(), 200);

pub fn body_bytes(self, bytes: impl AsRef<[u8]>) -> Self

Pass bytes as the request body.

Mime

The encoding is set to application/octet-stream.

Examples

let uri = "https://httpbin.org/post";
let data = b"hello world".to_owned();
let res = surf::post(uri).body_bytes(data).await?;
assert_eq!(res.status(), 200);

pub async fn body_file(self, path: impl AsRef<Path>) -> Result<Self>

Pass a file as the request body.

Mime

The encoding is set based on the file extension using mime_guess if the operation was successful. If path has no extension, or its extension has no known MIME type mapping, then None is returned.

Errors

This method will return an error if the file couldn’t be read.

Examples

let uri = "https://httpbin.org/post";
let res = surf::post(uri).body_file("./archive.tgz").await?.await?;
assert_eq!(res.status(), 200);

pub fn query(self, query: &impl Serialize) -> Result<Self, Error>

Set the URL querystring.

Examples

#[derive(Serialize, Deserialize)]
struct Index {
    page: u32
}

let query = Index { page: 2 };
let mut req = surf::get("https://httpbin.org/get").query(&query)?.build();
assert_eq!(req.url().query(), Some("page=2"));
assert_eq!(req.url().as_str(), "https://httpbin.org/get?page=2");

pub async fn recv_bytes(self) -> Result<Vec<u8>>

Submit the request and get the response body as bytes.

Examples

let bytes = surf::get("https://httpbin.org/get").recv_bytes().await?;
assert!(bytes.len() > 0);

pub async fn recv_string(self) -> Result<String>

Submit the request and get the response body as a string.

Examples

let string = surf::get("https://httpbin.org/get").recv_string().await?;
assert!(string.len() > 0);

pub async fn recv_json<T: DeserializeOwned>(self) -> Result<T>

Submit the request and decode the response body from json into a struct.

Examples

#[derive(Deserialize, Serialize)]
struct Ip {
    ip: String
}

let uri = "https://api.ipify.org?format=json";
let Ip { ip } = surf::get(uri).recv_json().await?;
assert!(ip.len() > 10);

pub async fn recv_form<T: DeserializeOwned>(self) -> Result<T>

Submit the request and decode the response body from form encoding into a struct.

Errors

Any I/O error encountered while reading the body is immediately returned as an Err.

If the body cannot be interpreted as valid json for the target type T, an Err is returned.

Examples

#[derive(Deserialize, Serialize)]
struct Body {
    apples: u32
}

let url = "https://api.example.com/v1/response";
let Body { apples } = surf::get(url).recv_form().await?;

pub fn middleware(self, middleware: impl Middleware) -> Self

Push middleware onto a per-request middleware stack.

Important: Setting per-request middleware incurs extra allocations. Creating a Client with middleware is recommended.

Client middleware is run before per-request middleware.

See the middleware submodule for more information on middleware.

Examples

let res = surf::get("https://httpbin.org/get")
    .middleware(surf::middleware::Redirect::default())
    .await?;

pub fn build(self) -> Request

Return the constructed Request.

pub async fn send(self) -> Result<Response>

Create a Client and send the constructed Request from it.